refactor(InputTest) : InputTest를 하는 객체를 분리
