+++ src/test/java/lotto/domain/LottoAnnouncementTest.java
@@ -3,7 +3,6 @@ package lotto.domain;
 import java.util.Arrays;
 import java.util.List;
 import lotto.exception.LottoAnnouncementException;
-import lotto.exception.LottoPiecesException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
