+++ src/test/java/lotto/utility/InputTest.java
@@ -1,0 +1,29 @@
+package lotto.utility;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.io.PrintStream;
+
+public class InputTest {
+    private final InputStream systemIn = System.in;
+    private final PrintStream systemOut = System.out;
+    private ByteArrayInputStream testIn;
+    private ByteArrayOutputStream testOut;
+
+    public void setUp() {
+        testOut = new ByteArrayOutputStream();
+        System.setOut(new PrintStream(testOut));
+    }
+
+    public void provideInput(String inputStringData) {
+        testIn = new ByteArrayInputStream(inputStringData.getBytes());
+        System.setIn(testIn);
+    }
+
+    public void restoreSystemInputOutput() {
+        System.setIn(systemIn);
+        System.setOut(systemOut);
+    }
+
+}
